--- a/appveyor.yml
+++ b/appveyor.yml
@@ -10,7 +10,7 @@ build_script:
   - dotnet --version
   - dotnet pack -c Release
 test_script:
-  - dotnet test .\test\Esprima.Tests\Esprima.Tests.csproj -c Release
+  - dotnet test .\test\Esprima.Tests\Esprima.Tests.csproj -c Release /p:Checked=true
 artifacts:
   - path: 'src\**\*.*nupkg'
 deploy: