--- a/src/Esprima/Loc.cs
+++ b/src/Esprima/Loc.cs
@@ -1,35 +1,37 @@
-﻿using System.Diagnostics;
-using static Esprima.EsprimaExceptionHelper;
+﻿using System;
+using System.Diagnostics;
 
 namespace Esprima
 {
-    using System;
-
     [DebuggerDisplay("{ToString()}")]
     public readonly struct Location : IEquatable<Location>
     {
         public Position Start  { get; }
         public Position End    { get; }
-        public string?   Source { get; }
+        public string?  Source { get; }
 
         public Location(Position start, Position end) : this(start, end, null)
         {
         }
 
-        public Location(Position start, Position end, string? source)
+        public Location(in Position start, in Position end, string? source)
         {
+#if LOCATION_ASSERTS
+            if (start == default && end != default
+                || end == default && start != default
+                || end.Line < start.Line
+                || start.Line > 0 && start.Line == end.Line && end.Column < start.Column)
+            {
+                EsprimaExceptionHelper.ThrowArgumentOutOfRangeException(nameof(end), end, Exception<ArgumentOutOfRangeException>.DefaultMessage);
+            }
+#endif
+            
             Start  = start;
-
-            End    = start == default && end != default
-                     || end == default && start != default
-                     || end.Line < start.Line
-                     || start.Line > 0 && start.Line == end.Line && end.Column < start.Column
-                   ? ThrowArgumentOutOfRangeException<Position>(nameof(end), end, Exception<ArgumentOutOfRangeException>.DefaultMessage)
-                   : end;
+            End = end;
             Source = source;
         }
 
-        public Location WithPosition(Position start, Position end) =>
+        public Location WithPosition(in Position start, in Position end) =>
             new Location(start, end, Source);
 
         public override bool Equals(object obj) =>
@@ -48,8 +50,8 @@ namespace Esprima
         public override string ToString() =>
             $"{Start}...{End}{(Source is string s ? ": " + s : null)}";
 
-        public static bool operator ==(Location left, Location right) => left.Equals(right);
-        public static bool operator !=(Location left, Location right) => !left.Equals(right);
+        public static bool operator ==(in Location left, in Location right) => left.Equals(right);
+        public static bool operator !=(in Location left, in Location right) => !left.Equals(right);
 
         public void Deconstruct(out Position start, out Position end)
         {