--- a/test/Esprima.Tests/ScannerTests.cs
+++ b/test/Esprima.Tests/ScannerTests.cs
@@ -8,7 +8,7 @@ namespace Esprima.Tests
         [Fact]
         public void CanScanMultiLineComment()
         {
-            var scanner = new Scanner("var foo=1; /* \"330413500\" */", new ParserOptions {Comment = true, Loc = true});
+            var scanner = new Scanner("var foo=1; /* \"330413500\" */", new ParserOptions {Comment = true});
             
             var results = new List<string>();
             Token token;