As these are struct to begin with, it's not a big penalty to include them always (and now it's actually faster). I tweaked the code a bit so only when testing we check the location line and column, JS version doesn't check them either. Also tweaked location passing to be using `in` to reduce some allocations. Removed the options from configuration as unneeded.

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|54.383 ms|1.0115 ms|3888.8889|1444.4444|555.5556|21.26 MB|
| **New** |	|	| **51.332 ms (-6%)** | **1.0094 ms** | **3600.0000 (-7%)** | **1400.0000 (-3%)** | **600.0000 (+8%)** | **20.45 MB (-4%)** |
| Old |ParseProgram|backbone-1.1.0|7.047 ms|0.0449 ms|617.1875|257.8125|109.3750|3.45 MB|
| **New** |	|	| **6.450 ms (-8%)** | **0.0221 ms** | **601.5625 (-3%)** | **242.1875 (-6%)** | **109.3750 (0%)** | **3.32 MB (-4%)** |
| Old |ParseProgram|jquery-1.9.1|46.479 ms|0.9223 ms|3363.6364|1272.7273|454.5455|18.13 MB|
| **New** |	|	| **44.180 ms (-5%)** | **0.8818 ms** | **3307.6923 (-2%)** | **1307.6923 (+3%)** | **538.4615 (+18%)** | **17.42 MB (-4%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|77.086 ms|1.5404 ms|5142.8571|2000.0000|857.1429|29.02 MB|
| **New** |	|	| **64.977 ms (-16%)** | **1.2618 ms** | **4857.1429 (-6%)** | **2142.8571 (+7%)** | **857.1429 (0%)** | **27.9 MB (-4%)** |
| Old |ParseProgram|mootools-1.4.5|39.090 ms|0.7786 ms|2642.8571|1071.4286|357.1429|14.81 MB|
| **New** |	|	| **34.171 ms (-13%)** | **0.0935 ms** | **2533.3333 (-4%)** | **1000.0000 (-7%)** | **466.6667 (+31%)** | **14.21 MB (-4%)** |
| Old |ParseProgram|underscore-1.5.2|3.931 ms|0.0035 ms|500.0000|210.9375|0.0000|2.85 MB|
| **New** |	|	| **3.569 ms (-9%)** | **0.0023 ms** | **496.0938 (-1%)** | **238.2813 (+13%)** | **0.0000** | **2.73 MB (-4%)** |
| Old |ParseProgram|yui-3.12.0|34.505 ms|0.6384 ms|2500.0000|1071.4286|357.1429|14.04 MB|
| **New** |	|	| **34.118 ms (-1%)** | **0.5066 ms** | **2357.1429 (-6%)** | **1000.0000 (-7%)** | **357.1429 (0%)** | **13.51 MB (-4%)** |